Joining queries in a new file
